[filegraph] Incorporate the file structure
There exist edge cases where the git-based distance does not work well.
For example, there are web tests that are added without modifying any
code-under-test. The biggest offender might be WPT dumps:
https://chromium-review.googlesource.com/q/owner:wpt-autoroller%2540chops-service-accounts.iam.gserviceaccount.com

Such edge cases drive the long-tail - influence the savings in high
ChangeRecall, e.g. 99% and 100%.

Introduce a new distance formula, based on the file structure: the
length of the path between two files in the file *tree*. For example,
files living in the same directory are more related than files living
in different directories. The key idea is the same: we utilize
developers' insight to group files together, but instead of grouping by
commits, this is grouping into directories.

Combine the two distances with a simple min() function: if one model
votes to select the test file, then we select it. This way, the new
distance compensates for the weaknesses of the other.

Before combining two distances, normalize them to use the same scale
by dividing by the maximum. Overall, evaluation now is done in 3 phases:
1. Evaluate only safety with only change-log-based distance and compute
   the max distance.
2. Do the same for file-structure-based distance.
3. Run the full evaluation with both distances combined, where each one
   is multiplied by 100/max.
   This scales both distances to the range [0, 100].

Before this CL:

  ChangeRecall | Savings | TestRecall | Distance
  ----------------------------------------------
  97.02%      |  65.75% |  52.64%    |  9.144
  98.03%      |  54.30% |  69.22%    | 10.890
  99.04%      |  31.19% |  87.20%    | 14.409
  100.00%     |   0.00% | 100.00%    |   +Inf

After this CL for the same data:

  ChangeRecall | Savings | TestRecall | Distance
  ----------------------------------------------
   97.02%      |  63.02% |  63.38%    | 30.642
   98.03%      |  56.89% |  71.18%    | 32.259
   99.10%      |  43.51% |  85.37%    | 37.500
  100.00%      |  16.10% |  97.35%    | 50.450

Bug: 1090668
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2681313
Auto-Submit: Nodir Turakulov <nodir@chromium.org>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#38499}

+++ go/src/infra/rts/presubmit/eval/eval_test.go
@@ -126,7 +126,7 @@ func TestQuantiles(t *testing.T) {
 				{Distance: 3},
 				{Distance: 4},
 			}
-			So(distanceQuantiles(afs, 2), ShouldResemble, []float64{2, 4})
+			So(distanceQuantiles(afs, 2), ShouldResemble, []float32{2, 4})
 		})
 		Convey(`4-quantiles of 1, 2, 3, 4`, func() {
 			afs := []rts.Affectedness{
@@ -135,7 +135,7 @@ func TestQuantiles(t *testing.T) {
 				{Distance: 3},
 				{Distance: 4},
 			}
-			So(distanceQuantiles(afs, 4), ShouldResemble, []float64{1, 2, 3, 4})
+			So(distanceQuantiles(afs, 4), ShouldResemble, []float32{1, 2, 3, 4})
 		})
 		Convey(`10-quantiles of 1, 2, 3, 4`, func() {
 			afs := []rts.Affectedness{
@@ -144,7 +144,7 @@ func TestQuantiles(t *testing.T) {
 				{Distance: 3},
 				{Distance: 4},
 			}
-			So(distanceQuantiles(afs, 10), ShouldResemble, []float64{1, 1, 2, 2, 2, 3, 3, 4, 4, 4})
+			So(distanceQuantiles(afs, 10), ShouldResemble, []float32{1, 1, 2, 2, 2, 3, 3, 4, 4, 4})
 		})
 	})
 }
